+++ lib/creator/widgets/background.dart
@@ -25,8 +25,7 @@ class BackgroundWidget extends CreatorWidget {
   /// Color of the page background
   Color color = Colors.white;
 
-  List<Color>? gradient;
-  BackgroundGradient gradientType = BackgroundGradient.type2;
+  CreativeGradient? gradient;
 
   BackgroundType type = BackgroundType.color;
 
@@ -54,24 +53,6 @@ class BackgroundWidget extends CreatorWidget {
         //     borderRadius: 15,
         //   ),
         // ),
-        Option.color(
-          palette: page.palette,
-          onChange: (color) {
-            if (color == null) return;
-            this.color = color;
-            if (type != BackgroundType.color) {
-              changeBackgroundType(BackgroundType.color);
-            }
-            updateListeners(WidgetChange.misc);
-          },
-          onChangeEnd: (color) {
-            if (color != null) this.color = color;
-            if (type != BackgroundType.color) {
-              changeBackgroundType(BackgroundType.color);
-            }
-            updateListeners(WidgetChange.update);
-          },
-        ),
         Option.button(
           icon: RenderIcons.palette,
           title: 'Palette',
@@ -98,47 +79,53 @@ class BackgroundWidget extends CreatorWidget {
             if (hasChanged) updateListeners(WidgetChange.update, historyMessage: 'Change Palette');
           },
         ),
-        Option.button(
-          icon: RenderIcons.resize,
-          title: 'Resize',
-          tooltip: 'Tap to resize the project',
-          onTap: (context) async {
-            EditorTab.modal(
-              context,
-              tab: (context, setState) => EditorTab.pickerBuilder(
-                title: 'Resize Project',
-                itemBuilder: (context, index) => Text('${PostSizePresets.values[index].title}'),
-                childCount: PostSizePresets.values.length,
-                onSelectedItemChanged: (index) {
-                  page.project.size = PostSizePresets.values[index].toSize();
-                  page.notifyListeners(PageChange.misc);
-                },
-              )
-            );
+        Option.color(
+          palette: page.palette,
+          onChange: (color) {
+            if (color == null) return;
+            this.color = color;
+            gradient = null;
+            if (type != BackgroundType.color) {
+              changeBackgroundType(BackgroundType.color);
+            }
+            updateListeners(WidgetChange.misc);
+          },
+          onChangeEnd: (color) {
+            if (color != null) this.color = color;
+            if (type != BackgroundType.color) {
+              changeBackgroundType(BackgroundType.color);
+            }
+            updateListeners(WidgetChange.update);
           },
         ),
         Option.button(
-          title: 'Padding',
+          title: 'Gradient',
           onTap: (context) async {
-            updateGrids(showGridLines: true, hideCenterGrids: true);
+            gradient ??= CreativeGradient.fromPalette(palette: page.palette);
+            updateListeners(WidgetChange.misc);
             await EditorTab.modal(
               context,
-              tab: (context, setState) => EditorTab.paddingEditor(
-                padding: padding,
-                max: page.project.contentSize.width/8,
-                min: 5,
-                onChange: (value) {
-                  padding = value;
-                  updateGrids(showGridLines: true, hideCenterGrids: true);
-                  updateListeners(WidgetChange.misc);
+              actions: [
+                IconButton(
+                  onPressed: () {
+                    gradient = null;
+                    Navigator.of(context).pop();
+                  },
+                  icon: Icon(RenderIcons.delete)
+                ),
+              ],
+              tab: (context, setState) => gradient!.getEditor(
+                palette: page.palette,
+                allowOpacity: false,
+                onChange: (change) {
+                  setState(() {});
+                  updateListeners(change);
                 },
-              )
+              ),
             );
-            updateGrids(showGridLines: false);
             updateListeners(WidgetChange.update);
           },
-          icon: RenderIcons.padding,
-          tooltip: 'Adjust Padding'
+          icon: RenderIcons.gradient
         ),
         Option.button(
           icon: RenderIcons.image,
@@ -194,6 +181,48 @@ class BackgroundWidget extends CreatorWidget {
             }
           },
         ),
+        Option.button(
+          icon: RenderIcons.resize,
+          title: 'Resize',
+          tooltip: 'Tap to resize the project',
+          onTap: (context) async {
+            EditorTab.modal(
+              context,
+              tab: (context, setState) => EditorTab.pickerBuilder(
+                title: 'Resize Project',
+                itemBuilder: (context, index) => Text('${PostSizePresets.values[index].title}'),
+                childCount: PostSizePresets.values.length,
+                onSelectedItemChanged: (index) {
+                  page.project.size = PostSizePresets.values[index].toSize();
+                  page.notifyListeners(PageChange.misc);
+                },
+              )
+            );
+          },
+        ),
+        Option.button(
+          title: 'Padding',
+          onTap: (context) async {
+            updateGrids(showGridLines: true, hideCenterGrids: true);
+            await EditorTab.modal(
+              context,
+              tab: (context, setState) => EditorTab.paddingEditor(
+                padding: padding,
+                max: page.project.contentSize.width/8,
+                min: 5,
+                onChange: (value) {
+                  padding = value;
+                  updateGrids(showGridLines: true, hideCenterGrids: true);
+                  updateListeners(WidgetChange.misc);
+                },
+              )
+            );
+            updateGrids(showGridLines: false);
+            updateListeners(WidgetChange.update);
+          },
+          icon: RenderIcons.padding,
+          tooltip: 'Adjust Padding'
+        ),
         if (page.project.pages.length > 1) Option.button(
           icon: RenderIcons.delete,
           title: 'Delete Page',
@@ -234,11 +263,7 @@ class BackgroundWidget extends CreatorWidget {
       child: Container(
         decoration: BoxDecoration(
           color: type == BackgroundType.color ? color : Colors.white,
-          gradient: (type == BackgroundType.gradient && gradient != null) ? LinearGradient(
-            colors: gradient!,
-            begin: gradientType.begin,
-            end: gradientType.end,
-          ) : null
+          gradient: gradient?.gradient
         ),
         child: (asset != null && imageProvider != null) ? imageProvider!.build(
           asset!,
@@ -363,37 +388,15 @@ class BackgroundWidget extends CreatorWidget {
     switch (_type) {
       case BackgroundType.color:
         asset = null;
-        gradient = null;
         imageProvider = null;
         break;
       case BackgroundType.image:
         gradient = null;
         break;
-      case BackgroundType.gradient:
-        asset = null;
-        imageProvider = null;
-        break;
       default:
     }
     type = _type;
     updateListeners(WidgetChange.misc);
-  }
-
-  List<String>? _generateGradientsHex() {
-    List<String> _generated = [];
-    if (gradient == null) return null;
-    for (Color color in gradient!) {
-      _generated.add(color.toHex());
-    }
-    return _generated;
-  }
-
-  List<Color> _generateGradientsColor(List<String> hex) {
-    List<Color> _generated = [];
-    for (String h in hex) {
-      _generated.add(HexColor.fromHex(h));
-    }
-    return _generated;
   }
 
   void onPaletteUpdate() {
@@ -413,7 +416,7 @@ class BackgroundWidget extends CreatorWidget {
     return {
       ... super.toJSON(buildInfo: buildInfo),
       'color': color.toHex(),
-      'gradient': _generateGradientsHex(),
+      'gradient': gradient?.toJSON(),
       'padding': padding.toJSON(),
       'image-provider': imageProvider?.toJSON(),
     };
@@ -429,10 +432,9 @@ class BackgroundWidget extends CreatorWidget {
       if (asset != null) {
         type = BackgroundType.image;
       }
-      if (json['gradient'] != null) {
-        gradient = _generateGradientsColor(json['gradient']);
-        type = BackgroundType.gradient;
-      }
+      if (json['gradient'] != null) try {
+        gradient = CreativeGradient.fromJSON(Map.from(json['gradient']));
+      } catch (e) { }
       if (json['padding'] != null) {
         padding = PaddingExtension.fromJSON(Map.from(json['padding']));
       }
@@ -452,47 +454,5 @@ class BackgroundWidget extends CreatorWidget {
 
 enum BackgroundType {
   color,
-  gradient,
   image
-}
-
-enum BackgroundGradient {
-  type1,
-  type2,
-  type3,
-  type4,
-}
-
-extension BackgroundGradientExtension on BackgroundGradient {
-
-  AlignmentGeometry get begin {
-    switch (this) {
-      case BackgroundGradient.type1:
-        return Alignment.centerLeft;
-      case BackgroundGradient.type2:
-        return Alignment.topLeft;
-      case BackgroundGradient.type3:
-        return Alignment.topCenter;
-      case BackgroundGradient.type4:
-        return Alignment.topRight;
-      default:
-        return Alignment.topLeft;
-    }
-  }
-
-  AlignmentGeometry get end {
-    switch (this) {
-      case BackgroundGradient.type1:
-        return Alignment.centerRight;
-      case BackgroundGradient.type2:
-        return Alignment.bottomRight;
-      case BackgroundGradient.type3:
-        return Alignment.bottomCenter;
-      case BackgroundGradient.type4:
-        return Alignment.bottomLeft;
-      default:
-        return Alignment.bottomRight;
-    }
-  }
-
 }
